+++ OccList.h
@@ -11,7 +11,7 @@ Header: OccList.h
 using namespace std;
 
 struct OccList { //the occurrence list for the FREQT, i.e., Asai's algorithm
-	vector<pair<int,short> > occurrence;
+	vector<pair<int,int> > occurrence;
 	int lastTid;
 	int mySupport;
 
@@ -19,15 +19,15 @@ struct OccList { //the occurrence list for the FREQT, i.e., Asai's algorithm
 	{
 	}
 
-	void insert(int& newTid, const short& newLocation);
+	void insert(int& newTid, const int& newLocation);
 
-	void insert(int& newTid, const short& newLocation, int& motherId);
+	void insert(int& newTid, const int& newLocation, int& motherId);
 
 	void explore(const vector<bool>& isFrequent, 
 		const vector<TextTree>& database,
 		const int& support,
 		vector<int>& frequency,
-		const short level);
+		const int level);
 };
 
 #endif //OccList_H
